+++ js/data.js
@@ -1,0 +1,22 @@
+import {randomNumberFromInterval} from './util.js'
+
+let data = [];
+let descriptions = [
+  'круто',
+  'классно',
+  'норм',
+  'нормас',
+  'прикольно'
+]
+
+for (let i = 1; i <= 25; i++) {
+  data.push({
+    id: i,
+    url: `photos/${i}.jpg`,
+    description: descriptions[randomNumberFromInterval(0, descriptions.length - 1)],
+    likes: randomNumberFromInterval(15, 200),
+    comments: randomNumberFromInterval(0, 200)
+  })
+}
+
+export {data};
